Controller PricesController tem metodo show onde traz o valor de acordo com o tempo estacinado
